Enhance script to support one off vs daemon mode

+++ DockerScripts/docker-build-and-run.sh
@@ -1,5 +1,4 @@
 #!/usr/bin/env bash
-set -i
 # Think of this as running make <target>
 # The prime image as en 'entrypoint' command of make and a default arg of 'test'
 # run this command with an argument like run, reset, etc as if this shel script
@@ -10,4 +9,9 @@ set -i
 
 docker-compose --file ../docker-compose.yml build --pull
 
-docker-compose --file ../docker-compose.yml run --name prime --rm --service-ports prime $1
+if [$1]
+    then
+        docker-compose --file ../docker-compose.yml run --name prime --rm --service-ports prime $1
+    else
+        docker-compose --file ../docker-compose.yml up
+fi
